cleanup: message consistency, DRY some code
ref #1944
