+++ package.nls.json
@@ -224,7 +224,6 @@
     "AWS.error.mfa.userCancelled": "User cancelled entering authentication code",
     "AWS.error.no.error.code": "No error code",
     "AWS.error.sam.local.package_json_not_found": "Unable to find package.json related to {0}",
-    "AWS.error.check.logs": "Check the logs for more information by running the \"{0}\" command from the {1}.",
     "AWS.explorerNode.addRegion": "Add a region to {0} Explorer...",
     "AWS.explorerNode.addRegion.tooltip": "Click here to add a region to {0} Explorer.",
     "AWS.explorerNode.apigateway.error": "Error loading API Gateway REST APIs",
@@ -466,7 +465,6 @@
     "AWS.cloudWatchLogs.limit.desc": "Maximum amount of log entries pulled per request from CloudWatch Logs (max 10000)",
     "AWS.samcli.detect.settings.updated": "Settings updated.",
     "AWS.samcli.detect.settings.not.updated": "No settings changes necessary.",
-    "AWS.samcli.deploy.general.error": "An error occurred while deploying a SAM Application. {0}",
     "AWS.samcli.deploy.workflow.init": "Building SAM Application...",
     "AWS.samcli.deploy.workflow.packaging": "Packaging SAM Application to S3 Bucket: {0}",
     "AWS.samcli.deploy.workflow.packaging.noBuild": "Attempting to package source template directory directly after \"sam build\" failed",
@@ -511,8 +509,6 @@
     "AWS.samcli.userChoice.browse": "Locate SAM CLI...",
     "AWS.samcli.userChoice.visit.install.url": "Get SAM CLI",
     "AWS.samcli.userChoice.update.awstoolkit.url": "Visit Marketplace",
-    "AWS.samcli.initWizard.general.error": "An error occurred while creating a new SAM Application. {0}",
-    "AWS.samcli.initWizard.resume.error": "An error occured while resuming SAM Application creation. {0}",
     "AWS.wizard.location.prompt": "Select a workspace folder for your new project",
     "AWS.initWizard.location.select.folder": "Select a different folder...",
     "AWS.initWizard.location.select.folder.empty.workspace": "There are no workspace folders open. Select a folder...",
